Fixed plugin loading issue due to LP compatibility, Fixed Logo not showing up
