+++ NewMod/NewMod.csproj
@@ -15,13 +15,13 @@
 	</PropertyGroup>
 
 	<PropertyGroup Condition=" '$(Configuration)' == 'ANDROID' ">
-	    <RestoreSources>$(RestoreSources);$(MSBuildProjectDirectory)\libs\Android</RestoreSources>
+	    <RestoreSources>$(RestoreSources);$(MSBuildProjectDirectory)\..\libs\Android</RestoreSources>
 		<DefineConstants>TRACE;ANDROID_BUILD</DefineConstants>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<PackageReference Include="Reactor" Version="2.3.1" />
-		<PackageReference Include="AllOfUs.MiraAPI" Version="0.2.0-rc.6" />
+		<PackageReference Include="AllOfUs.MiraAPI" Version="0.2.0-ci.558" />
 		<PackageReference Condition="'$(Configuration)' == 'Debug' Or '$(Configuration)' == 'Release' " Include="AmongUs.GameLibs.Steam" Version="2025.4.15" PrivateAssets="all" />
 		<PackageReference Condition="'$(Configuration)' == 'ANDROID' " Include="AmongUs.GameLibs.Android" Version="2024.10.29" PrivateAssets="all" />
 		<PackageReference Include="BepInEx.AutoPlugin" Version="1.1.0" PrivateAssets="all" />
