+++ nuget.config
@@ -3,5 +3,6 @@
 <configuration>
     <packageSources>
         <add key="BepInEx" value="https://nuget.bepinex.dev/v3/index.json" />
+        <add key="MiraAPI" value="libs/" />
     </packageSources>
 </configuration>
